Moved Handler to the separate file.
